+++ Products/CMFPlone/utils.py
@@ -16,6 +16,7 @@ from log import log_exc
 from os.path import join, abspath, split
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from plone.registry.interfaces import IRegistry
+from plone.subrequest.interfaces import ISubRequest
 from Products.CMFCore.permissions import ManageUsers
 from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.utils import ToolInit as CMFCoreToolInit
@@ -716,3 +717,13 @@ def get_installer(context, request=None):
         request = aq_get(context, 'REQUEST', None)
     view = getMultiAdapter((context, request), name='installer')
     return view
+
+
+def get_top_request(request):
+    """Get highest request from a subrequest.
+    """
+    def _top_request(req):
+        if ISubRequest.providedBy(req):
+            return _top_request(req.get('PARENT_REQUEST', None))
+        return req
+    return _top_request(request)